feat: admin

+++ dubhe-data/src/main/java/org/dubhe/data/machine/state/specific/data/StrengtheningState.java
@@ -1,19 +1,4 @@
-/**
- * Copyright 2020 Tianshu AI Platform. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * =============================================================
- */
+
 package org.dubhe.data.machine.state.specific.data;
 
 import org.dubhe.biz.log.enums.LogEnum;
@@ -53,7 +38,7 @@ public class StrengtheningState extends AbstractDataState {
      * @param primaryKeyId 业务ID
      */
     @Override
-    public void strengtheningCompleteEvent(Integer primaryKeyId) {
+    public void strengtheningCompleteEvent(Long primaryKeyId) {
         LogUtil.debug(LogEnum.STATE_MACHINE, " 【增强中】 执行事件前内存中状态机的状态 :{} ", dataStateMachine.getMemoryDataState());
         LogUtil.debug(LogEnum.STATE_MACHINE, " 接受参数： {} ", primaryKeyId);
         datasetMapper.updateStatus(Long.valueOf(primaryKeyId), DataStateEnum.ANNOTATION_COMPLETE_STATE.getCode());
@@ -67,7 +52,7 @@ public class StrengtheningState extends AbstractDataState {
      * @param primaryKeyId 业务ID
      */
     @Override
-    public void strengtheningAutoCompleteEvent(Integer primaryKeyId) {
+    public void strengtheningAutoCompleteEvent(Long primaryKeyId) {
         LogUtil.debug(LogEnum.STATE_MACHINE, " 【增强中】 执行事件前内存中状态机的状态 :{} ", dataStateMachine.getMemoryDataState());
         LogUtil.debug(LogEnum.STATE_MACHINE, " 接受参数： {} ", primaryKeyId);
         datasetMapper.updateStatus(Long.valueOf(primaryKeyId), DataStateEnum.AUTO_TAG_COMPLETE_STATE.getCode());
